+++ test/test_scratchpad.py
@@ -123,3 +123,32 @@ def test_load_multiple_formats():
     assert lat == "N3747332"
     assert lon == "W11425693"
     assert alt == "4913"
+
+def test_load_steerpoint_without_altitude_in_meters():
+    """
+    GIVEN A DCS Scratchpad datafile with steerpoint where altitude only in feet
+    WHEN the class is instantiated and method is called
+    THEN load steerpoint DDM coordinates and elevation in feet
+    """
+
+    dataloader = DataLoader("test/resources/TestData-Scratchpad-No-Meter.txt")
+    dataloader.load_data()
+
+    waypoints = dataloader.get_waypoints()
+    assert waypoints is not None
+    assert type(waypoints) is dict
+    assert len(waypoints) == 1
+
+    assert waypoints[1] is not None
+
+    wp = waypoints[1]
+    assert wp is not None
+
+    lat, lon, alt = wp
+    assert lat is not None
+    assert lon is not None
+    assert alt is not None
+
+    assert lat == "N3747332"
+    assert lon == "W11425693"
+    assert alt == "152"